requirements.txt
einops>=0.6.1
numpy>=1.24.4
open-clip-torch>=2.20.0
opencv-python==4.6.0.66
pillow>=9.5.0
scipy==1.11.1
timm>=0.9.2
tokenizers

torchdata==0.6.1
torchmetrics>=1.0.1


tqdm>=4.65.0
transformers==4.35.2
triton==2.0.0
urllib3<1.27,>=1.25.4
xformers>=0.0.20
streamlit-keyup==0.2.0
lpips
peft==0.10.0
pyiqa
omegaconf
dominate
diffusers==0.25.1
gradio==3.43.1
huggingface_hub==0.22.2
